+++ adapter/http/router.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"net/http"
 
-	"github.com/yoshikawa-river/ChatApp/infla"
+	"github.com/yoshikawa-river/ChatApp/infra"
 	"github.com/yoshikawa-river/ChatApp/usecase"
 )
 
@@ -15,7 +15,7 @@ func InitRouter(db_info *sql.DB) *http.ServeMux {
 	router.HandleFunc("/server_health_check", ServerHealthCheck)
 	router.HandleFunc("/db_health_check", DBHealthCheck)
 
-	userRepository := infla.NewUserRepository(db_info)
+	userRepository := infra.NewUserRepository(db_info)
 	userUsecase := usecase.NewUserInteractor(userRepository)
 	userHandler := NewUserHandler(userUsecase)
