+++ infra/connect_db.go
@@ -1,4 +1,4 @@
-package infla
+package infra
 
 import (
 	"database/sql"
